Update test cluster
